+++ src/gc/state.h
@@ -52,8 +52,8 @@ typedef struct gc_state {
     size_t          collections_minor;
     size_t          collections_major;
     size_t          total_freed;
-    size_t          total_ns;       /* cumulative ns spent in minor and major collections */
-    size_t          max_ns;         /* largest single-collection ns */
+    uint64_t        total_ns;       /* cumulative ns spent in minor and major collections */
+    uint64_t        max_ns;         /* largest single-collection ns */
 
     /* Generational bookkeeping: bytes_young + bytes_old == bytes_alloc.
      * old_baseline captures bytes_old right after the last major sweep;
